Refactored Parceler specific helper and some other minor updates.
