+++ AndroidAnnotations/androidannotations-core/androidannotations-test/pom.xml
@@ -33,6 +33,7 @@
 		<main.basedir>${project.parent.parent.basedir}</main.basedir>
 		<android-rome-version>1.0.0-r2</android-rome-version>
 		<checkstyle.excludes>**/R.java,**/BuildConfig.java,</checkstyle.excludes>
+		<parceler.version>1.0.3</parceler.version>
 	</properties>
 
 	<dependencies>
@@ -85,13 +86,13 @@
 		<dependency>
 			<groupId>org.parceler</groupId>
 			<artifactId>parceler</artifactId>
-			<version>1.0.3</version>
+			<version>${parceler.version}</version>
 			<scope>provided</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.parceler</groupId>
 			<artifactId>parceler-api</artifactId>
-			<version>1.0.3</version>
+			<version>${parceler.version}</version>
 		</dependency>
 	</dependencies>
 	<build>
